almeval/models/funaudio/requirements.txt
# Basic Dependencies
transformers==4.52.3
datasets==3.6.0
numpy<2.0.0
accelerate
deepspeed==0.17.3
librosa

# CV3
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/ # https://github.com/microsoft/onnxruntime/issues/21684
conformer==0.3.2
diffusers==0.29.0
gdown==5.1.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
inflect==7.3.1
lightning==2.2.4
loguru
matplotlib==3.7.5
modelscope
networkx==3.1
omegaconf==2.3.0
onnx==1.16.0
onnxruntime-gpu==1.18.0; sys_platform == 'linux'
onnxruntime==1.18.0; sys_platform == 'darwin' or sys_platform == 'win32'
openai-whisper
protobuf==4.25
pyarrow==18.1.0
pydantic==2.7.0
pyworld==0.3.4
rich==13.7.1
soundfile==0.12.1
tensorrt-cu12==10.13.3.9; sys_platform == 'linux'
tensorrt-cu12-bindings==10.13.3.9; sys_platform == 'linux'
tensorrt-cu12-libs==10.13.3.9; sys_platform == 'linux'
uvicorn==0.30.0
wetext==0.0.4
wget==3.2
x_transformers